Changed afspraak bedrag to int

+++ huishoudboekje-service/migrations/versions/41b64919bfb4_added_afspraken_model.py
@@ -27,7 +27,7 @@ def upgrade():
     sa.Column('aantal_betalingen', sa.Integer(), nullable=True),
     sa.Column('interval', sa.String(), nullable=True),
     sa.Column('tegen_rekening_id', sa.Integer(), nullable=True),
-    sa.Column('bedrag', sa.Float(), nullable=True),
+    sa.Column('bedrag', sa.Integer(), nullable=True),
     sa.Column('credit', sa.Boolean(), nullable=True),
     sa.Column('kenmerk', sa.String(), nullable=True),
     sa.Column('actief', sa.Boolean(), nullable=True),
